improved speed and mem access

+++ dreamoxide-core/src/cpu.rs
@@ -14,8 +14,8 @@ use std::ops::IndexMut;
 pub const FPSCR_MASK : u32 = 0x003FFFFF;
 
 pub struct Cpu {
-    pub pc: u32,
-    pub pr: u32,
+    pub pc: usize,
+    pub pr: usize,
     pub status: StatusRegister,
     pub registers: [GeneralRegister; 24],
     pub fpu_registers: [FloatingPointRegister; 32],
@@ -28,7 +28,7 @@ pub struct Cpu {
     pub spc: GeneralRegister,
     pub fpscr: GeneralRegister,
     pub fpul: GeneralRegister,
-    pub max: u32,
+    pub max: usize,
 }
 
 impl Cpu {
@@ -52,6 +52,7 @@ impl Cpu {
         }
     }
 
+    #[inline(always)]
     pub fn step(&mut self, mem: &mut Memory) {
         if self.pc > self.max && self.pc < 0xa0000000 {
             self.max = self.pc;
@@ -62,11 +63,8 @@ impl Cpu {
                 self.pc += 2;
             },
             &MemoryField::MemoryCell(val) => {
-                //if val == 0xFFFF {
-                    //return;
-                //}
                 let inst = InstructionDecoder::decode(val);
-                *mem.access_mut(self.pc as usize) = MemoryField::InstructionCell(inst);
+                *mem.access_mut(self.pc) = MemoryField::InstructionCell(inst);
 
                 if inst == Instruction::Unknown {
                     println!("[0x{:08x}] Could not decode {:04x}", self.pc, val);
@@ -78,16 +76,19 @@ impl Cpu {
         }
     }
 
+    #[inline(always)]
     pub fn fpu<'a>(&'a self, reg: Operand) -> &'a FloatingPointRegister {
         let bank = if self.fpscr.value & 0x200000 != 0 { 16 } else { 0 };
         &self.fpu_registers[bank + reg.unwrap() as usize]
     }
 
+    #[inline(always)]
     pub fn fpu_mut<'a>(&'a mut self, reg: Operand) -> &'a mut FloatingPointRegister {
         let bank = if self.fpscr.value & 0x200000 != 0 { 16 } else { 0 };
         &mut self.fpu_registers[bank + reg.unwrap() as usize]
     }
 
+    #[inline(always)]
     pub fn banked<'a>(&'a self, reg: Operand) -> &'a GeneralRegister {
         let bank = if self.status.is_banked() && self.status.is_privileged() && reg.unwrap() < 8 { 0 } else { 16 };
         &self.registers[bank + reg.unwrap() as usize]
@@ -98,6 +99,7 @@ impl Cpu {
 impl Index<Operand> for Cpu {
     type Output = GeneralRegister;
 
+    #[inline(always)]
     fn index<'a>(&'a self, _index: Operand) -> &'a GeneralRegister {
         let bank = if self.status.is_banked() && self.status.is_privileged() && _index.unwrap() < 8 { 16 } else { 0 };
         &self.registers[bank + _index.unwrap() as usize]
@@ -105,6 +107,7 @@ impl Index<Operand> for Cpu {
 }
 
 impl IndexMut<Operand> for Cpu {
+    #[inline(always)]
     fn index_mut<'a>(&'a mut self, _index: Operand) -> &'a mut GeneralRegister {
         let bank = if self.status.is_banked() && self.status.is_privileged() && _index.unwrap() < 8 { 16 } else { 0 };
         &mut self.registers[bank + _index.unwrap() as usize]
